+++ core/src/main/java/org/elasticsearch/search/aggregations/bucket/significant/SignificantStringTerms.java
@@ -103,15 +103,8 @@ public class SignificantStringTerms extends InternalMappedSignificantTerms<Signi
         }
 
         @Override
-        public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-            builder.startObject();
-            builder.field(CommonFields.KEY.getPreferredName(), getKeyAsString());
-            builder.field(CommonFields.DOC_COUNT.getPreferredName(), getDocCount());
-            builder.field("score", score);
-            builder.field("bg_count", supersetDf);
-            aggregations.toXContentInternal(builder, params);
-            builder.endObject();
-            return builder;
+        protected XContentBuilder keyToXContent(XContentBuilder builder) throws IOException {
+            return builder.field(CommonFields.KEY.getPreferredName(), getKeyAsString());
         }
 
         @Override
@@ -160,21 +153,6 @@ public class SignificantStringTerms extends InternalMappedSignificantTerms<Signi
     protected SignificantStringTerms create(long subsetSize, long supersetSize, List<Bucket> buckets) {
         return new SignificantStringTerms(getName(), requiredSize, minDocCount, pipelineAggregators(), getMetaData(), format, subsetSize,
                 supersetSize, significanceHeuristic, buckets);
-    }
-
-    @Override
-    public XContentBuilder doXContentBody(XContentBuilder builder, Params params) throws IOException {
-        builder.field("doc_count", subsetSize);
-        builder.startArray(CommonFields.BUCKETS.getPreferredName());
-        for (Bucket bucket : buckets) {
-            //There is a condition (presumably when only one shard has a bucket?) where reduce is not called
-            // and I end up with buckets that contravene the user's min_doc_count criteria in my reducer
-            if (bucket.subsetDf >= minDocCount) {
-                bucket.toXContent(builder, params);
-            }
-        }
-        builder.endArray();
-        return builder;
     }
 
     @Override
